+++ ecd-20200930/alibabacloud_ecd20200930/models/_modify_timer_group_request.py
@@ -252,6 +252,8 @@ class ModifyTimerGroupRequestConfigTimersSegmentTimers(DaraModel):
         timer_order: int = None,
         timezone: str = None,
         trigger_type: str = None,
+        verification_notification_time: int = None,
+        verification_time: int = None,
     ):
         self.appointment_timer = appointment_timer
         self.end_cron_expression = end_cron_expression
@@ -267,6 +269,8 @@ class ModifyTimerGroupRequestConfigTimersSegmentTimers(DaraModel):
         self.timer_order = timer_order
         self.timezone = timezone
         self.trigger_type = trigger_type
+        self.verification_notification_time = verification_notification_time
+        self.verification_time = verification_time
 
     def validate(self):
         pass
@@ -318,6 +322,12 @@ class ModifyTimerGroupRequestConfigTimersSegmentTimers(DaraModel):
         if self.trigger_type is not None:
             result['TriggerType'] = self.trigger_type
 
+        if self.verification_notification_time is not None:
+            result['VerificationNotificationTime'] = self.verification_notification_time
+
+        if self.verification_time is not None:
+            result['VerificationTime'] = self.verification_time
+
         return result
 
     def from_map(self, m: dict = None):
@@ -363,6 +373,12 @@ class ModifyTimerGroupRequestConfigTimersSegmentTimers(DaraModel):
 
         if m.get('TriggerType') is not None:
             self.trigger_type = m.get('TriggerType')
+
+        if m.get('VerificationNotificationTime') is not None:
+            self.verification_notification_time = m.get('VerificationNotificationTime')
+
+        if m.get('VerificationTime') is not None:
+            self.verification_time = m.get('VerificationTime')
 
         return self